agentes/04_pauta/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.28.0
google-api-python-client>=2.120.0
